moved some files to legacy and implemented redoc for api documentation
