+++ api/versionController.js
@@ -1,11 +1,11 @@
 var express = require("express");
 var router = express.Router();
 
-var version1Router = require('./v1/v1');
+var version1Router = require('./v1/router');
 var generateKey = require('./generateKey');
 
 router.get('/', function(req, res) {
-    res.redirect('/api/v1');
+    res.redirect('/v1');
 });
 
 router.use("/v1", version1Router);
