+++ email/apiConfirmation/html.ejs
@@ -183,7 +183,7 @@
                                                     <a
 													href="<%= confirmationURL %>"
 													style="-webkit-text-size-adjust: none; text-decoration: none; display: inline-block; color: #ffffff;     background: #2484c6;
-                                                    background: radial-gradient(circle, #2484C6 0%, #74D1F6 100%); border-radius: 4px; -webkit-border-radius: 4px; -moz-border-radius: 4px; width: auto; width: auto; border-top: 1px solid #3AAEE0; border-right: 1px solid #3AAEE0; border-bottom: 1px solid #3AAEE0; border-left: 1px solid #3AAEE0; padding-top: 5px; padding-bottom: 5px; font-family: Arial, Helvetica Neue, Helvetica, sans-serif; text-align: center; mso-border-alt: none; word-break: keep-all;"
+                                                     border-radius: 4px; -webkit-border-radius: 4px; -moz-border-radius: 4px; width: auto; width: auto; padding-top: 5px; padding-bottom: 5px; font-family: Arial, Helvetica Neue, Helvetica, sans-serif; text-align: center; mso-border-alt: none; word-break: keep-all;"
 													>
                                                     <span
                                                         style="padding-left:20px;padding-right:20px;font-size:16px;display:inline-block;"><span
