+++ email/apiKeyDelivery/subject.ejs
@@ -1,0 +1,1 @@
+Thank you for verifying your PeterPortal API Key (<%= appName %>)
